requirements.txt
flask==3.1.*
sqlalchemy==2.0.*
pytest==8.*
pdfplumber
python-dotenv
plaid-python
